feat: huge refactor incoming
