+++ apps/backend/src/rooms/index.ts
@@ -1,1 +1,4 @@
-export { roomRouter } from "./room.routes";
+export { createRoomRouter } from "./room.routes";
+export { setUpRoomSocketListeners, type SocketState } from "./room.sockets";
+export { RoomStore } from "./room.store";
+export { RestaurantService } from "./restaurant.service";
